Extend Rval to include indication of contents being fully expanded
Currently, this is only valid for scalars, and I *think* that that's the only
context where it really makes sense.

I'm not sure that I'm happy with this change, as it makes creating Rval's more
error-prone, and it doesn't necessarily make sense to keep the status of an
Rval's expansion as part of it's body. Playing around with this idea.